refactor : 예외 처리 구분
